Transferred validate_insert decorator to new decorator file
